+++ gitoxide-core/src/repository.rs
@@ -3,5 +3,6 @@ use std::path::PathBuf;
 use anyhow::{Context as AnyhowContext, Result};
 
 pub fn init(directory: Option<PathBuf>) -> Result<git_repository::Path> {
-    git_repository::init::into(directory.unwrap_or_default()).with_context(|| "Repository initialization failed")
+    git_repository::path::create::into(directory.unwrap_or_default())
+        .with_context(|| "Repository initialization failed")
 }